+++ l10n/gittyup_es.ts
@@ -5932,89 +5932,89 @@ The submodule will be removed from this project. Its working files and cached lo
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="714"/>
+        <location filename="../src/ui/SideBar.cpp" line="711"/>
         <source>Close</source>
         <translation>Cerrar</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="718"/>
-        <location filename="../src/ui/SideBar.cpp" line="822"/>
-        <location filename="../src/ui/SideBar.cpp" line="917"/>
+        <location filename="../src/ui/SideBar.cpp" line="715"/>
+        <location filename="../src/ui/SideBar.cpp" line="819"/>
+        <location filename="../src/ui/SideBar.cpp" line="914"/>
         <source>Remove</source>
         <translation>Remover</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="723"/>
+        <location filename="../src/ui/SideBar.cpp" line="720"/>
         <source>Authorize</source>
         <translation>Autorizar</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="738"/>
+        <location filename="../src/ui/SideBar.cpp" line="735"/>
         <source>Clone Repository</source>
         <translation>Clonar Repositorio</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="749"/>
+        <location filename="../src/ui/SideBar.cpp" line="746"/>
         <source>Open Existing Repository</source>
         <translation>Abrir Repositorio Existente</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="753"/>
+        <location filename="../src/ui/SideBar.cpp" line="750"/>
         <source>Open Repository</source>
         <translation>Abrir Repositorio</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="762"/>
+        <location filename="../src/ui/SideBar.cpp" line="759"/>
         <source>Initialize New Repository</source>
         <translation>Inicializar Nuevo Repositorio</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="777"/>
+        <location filename="../src/ui/SideBar.cpp" line="774"/>
         <source>Add %1 Account</source>
         <translation>Añadir cuenta de %1</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="812"/>
+        <location filename="../src/ui/SideBar.cpp" line="809"/>
         <source>&lt;p&gt;Are you sure you want to remove the remote repository association for %1?&lt;/p&gt;&lt;p&gt;The local clone itself will not be affected.&lt;/p&gt;</source>
         <translation>&lt;p&gt;¿Estás seguro que deseas remover la asociación al repositorio remoto de %1?&lt;/p&gt;&lt;p&gt;El repositorio local no será afectado.&lt;/p&gt;</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="817"/>
+        <location filename="../src/ui/SideBar.cpp" line="814"/>
         <source>Remove Repository Association?</source>
         <translation>¿Remover Asociación de Repositorio?</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="841"/>
+        <location filename="../src/ui/SideBar.cpp" line="838"/>
         <source>Clear All Recent</source>
         <translation>Limpiar Todos los Recientes</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="845"/>
+        <location filename="../src/ui/SideBar.cpp" line="842"/>
         <source>Show Full Path</source>
         <translation>Mostrar Ruta Completa</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="855"/>
+        <location filename="../src/ui/SideBar.cpp" line="852"/>
         <source>Filter Non-existent Paths</source>
         <translation>Filtrar Rutas no Existentes</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="863"/>
+        <location filename="../src/ui/SideBar.cpp" line="860"/>
         <source>Refresh Remote Accounts</source>
         <translation>Refrescar Cuentas Remotas</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="870"/>
+        <location filename="../src/ui/SideBar.cpp" line="867"/>
         <source>Show Full Name</source>
         <translation>Mostrar Nombre Completo</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="906"/>
+        <location filename="../src/ui/SideBar.cpp" line="903"/>
         <source>&lt;p&gt;Are you sure you want to remove the %1 account for &apos;%2&apos;?&lt;/p&gt;&lt;p&gt;Only the account association will be removed. Remote configurations and local clones will not be affected.&lt;/p&gt;</source>
         <translation>&lt;p&gt;¿Estás seguro que deseas remover la cuenta de %1 para &apos;%2&apos;?&lt;/p&gt;&lt;p&gt;Solo la asociación de la cuenta será removida. La configuraciones remotas y repositorios locales no serán afectados.&lt;/p&gt;</translation>
     </message>
     <message>
-        <location filename="../src/ui/SideBar.cpp" line="911"/>
+        <location filename="../src/ui/SideBar.cpp" line="908"/>
         <source>Remove Account?</source>
         <translation>¿Remover Cuenta?</translation>
     </message>